third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-position-y-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['background-position-y'] = "center" should set the property value
  assert_equals: serialization should be canonical expected "center" but got "50%"
[FAIL] e.style['background-position-y'] = "top" should set the property value
  assert_equals: serialization should be canonical expected "top" but got "0%"
[FAIL] e.style['background-position-y'] = "bottom" should set the property value
  assert_equals: serialization should be canonical expected "bottom" but got "100%"
[FAIL] e.style['background-position-y'] = "y-start" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-position-y'] = "y-end" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-position-y'] = "top -20%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-position-y'] = "bottom 10px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-position-y'] = "center, top, bottom" should set the property value
  assert_equals: serialization should be canonical expected "center, top, bottom" but got "50%, 0%, 100%"
[FAIL] e.style['background-position-y'] = "0.5em, y-start, y-end" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['background-position-y'] = "calc(10px - 0.5em), top -20%, bottom 10px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
